requirements.txt
python-dotenv
discord.py
asynciopip 
requests
